Revert "cheri: Fix elf/tst-dlmodcount test"
This reverts commit 3a69d7f241ab9f738479e039bd6eecf433868375.

+++ elf/tst-dlmodcount.c
@@ -30,7 +30,7 @@ static int
 callback (struct dl_phdr_info *info, size_t size, void *ptr)
 {
   static int last_adds = 0, last_subs = 0;
-  unsigned long cmd = (intptr_t) ptr;
+  intptr_t cmd = (intptr_t) ptr;
 
   printf ("  size = %Zu\n", size);
   if (size < (offsetof (struct dl_phdr_info, dlpi_subs)
